controlador de titulados
